+++ playground/legacy-evm-react-vite-playground/vite.config.ts
@@ -1,6 +1,5 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
-import { nodePolyfills } from 'vite-plugin-node-polyfills';
 
 // https://vite.dev/config/
 export default defineConfig({
@@ -14,6 +13,6 @@ export default defineConfig({
     outDir: './build',
     emptyOutDir: true, // also necessary
   },
-  plugins: [react(), nodePolyfills({ include: ['buffer'] })],
+  plugins: [react()],
   base: './',
 });